processing/requirements.txt
# Function dependencies, for example:
# google-cloud-firestore
google-cloud-firestore==2.16.0